+++ noxfiles/ci_nox.py
@@ -1,4 +1,6 @@
 """Contains the nox sessions used during CI checks."""
+from time import sleep
+
 import nox
 from constants_nox import (
     CI_ARGS,
@@ -80,23 +82,25 @@ def pylint(session: nox.Session) -> None:
     if session.posargs == ["docker"]:
         run_command = (*RUN_STATIC_ANALYSIS, "pylint")
     else:
-        run_command = ("pylint", "src", "noxfiles")
+        run_command = ("pylint", "src", "noxfiles", "tests")
     session.run(*run_command, external=True)
 
 
 @nox.session()
 def xenon(session: nox.Session) -> None:
-    """Run the 'pylint' code linter."""
+    """Run 'xenon' code complexity monitoring."""
     if session.posargs == ["docker"]:
         run_command = (*RUN_STATIC_ANALYSIS, "xenon")
     else:
         run_command = (
             "xenon",
+            "noxfiles",
             "src",
+            "tests",
             "--max-absolute B",
             "--max-modules B",
             "--max-average A",
-            "--ignore 'data, tests, docs'",
+            "--ignore 'data, docs'",
             "--exclude src/fidesctl/_version.py",
         )
     session.run(*run_command, external=True)
@@ -128,6 +132,7 @@ def fidesctl_db_scan(session: nox.Session) -> None:
     """Scan the fidesctl application database to check for dataset discrepancies."""
     session.notify("teardown")
     session.run(*START_APP, external=True)
+    sleep(10)
     run_command = (
         *RUN,
         "fidesctl",